# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Geolocal",
        "version": "1.0.0"
    },
    "paths": {
        "/api/arrondissements/": {
            "get": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "les differents arrondissements couvrir par l'api",
                "description": "Les arrondissements sont liées au commune",
                "operationId": "",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "",
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "koda"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/arrondissements": {
            "post": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "liste des problemes",
                "description": "",
                "operationId": "",
                "requestBody": {
                    "description": "Vous devez vous connectez d'abord et etre admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "commune_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "bohicon"
                                    },
                                    "commune_id": {
                                        "type": "int",
                                        "format": "int",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/arrondissements/{id}": {
            "delete": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "Suppression",
                "description": "suppression d'un arrondissement",
                "operationId": "",
                "requestBody": {
                    "description": "Uniquement par un compte admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Arrondissements"
                ],
                "summary": "modificatins",
                "description": "Pour modifier un attributs des problemes vous devez etre admin",
                "operationId": "",
                "requestBody": {
                    "description": "Vous devez vous connectez d'abord et etre admin ",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}